apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "A LangGraph based backend to orchestrate distributed MCP servers for lab machines"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langgraph>=1.0.6",
    "langchain>=1.2.3",
    "langchain-openai>=1.0.0",
    "langsmith>=0.6.2",
    "python-dotenv==1.2.1",
    "ddgs>=9.10.0",
]
